move LinkageChecker to standalone file
